finished exercise 1

+++ Exercício #1 - C#/Exercicio1/Exercicio1.Library/Models/AverageCSV.cs
@@ -8,6 +8,6 @@ namespace Exercicio1.Library.Models
     {
         public string Name { get; set; }
         public string City { get; set; }
-        public float Age { get; set; }
+        public decimal Age { get; set; }
     }
 }
